 \docType{data}
    \name{stockholm}
    \alias{stockholm}
    \title{The Stockholm Cohort from Ross-Adams et al. (2015)}
    \format{a \code{ExpressionSet} instance, 1 row per Illumina Huamnv4 probe.}
    \source{
      GSE70769
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO. Details in vignette.
    }
        \value{ExpressionSet object for the dataset}
    \author{
Mark Dunning
    }
    \usage{camcap}
    \keyword{datasets}